not important

+++ src/training/learner/learners/learners.py
@@ -88,7 +88,7 @@ class BasicLearner(Learner):
         """
         # Image
         if not self._params["reduce_dimensions"]:
-            batch["image"]: torch.Tensor = batch["image"].type(torch.float32).to(self._DEVICE)
+            batch["image"]: torch.Tensor = batch["image"].to(self._DEVICE)
 
         # Predicts added noise
         noisy_image, noise, timestep = self._add_noise(batch["image"])
